+++ Crypto/Infrastructure/Services/AddressService.cs
@@ -17,7 +17,7 @@ namespace Binebase.Exchange.CryptoService.Infrastructure.Services
         public AddressService(IOptions<Configuration> options)
             => _configuration = options.Value;
 
-        public async Task<string> GenerateAddress(Currency currency, int index)
+        public async Task<string> GenerateAddress(Currency currency, uint index)
             => currency switch
             {
                 Currency.BTC => await GenerateBitcoinAddress(index),
@@ -25,18 +25,18 @@ namespace Binebase.Exchange.CryptoService.Infrastructure.Services
                 _ => throw new NotSupportedException(),
             };
 
-        private async Task<string> GenerateBitcoinAddress(int index)
+        private async Task<string> GenerateBitcoinAddress(uint index)
         {
             var mnemo = new Mnemonic(_configuration.Mnemonic, Wordlist.English);
             var key = mnemo.DeriveExtKey(_configuration.Passwords[Currency.BTC]);
-            var address = key.Derive((uint)index).ScriptPubKey.GetDestinationAddress(_configuration.IsTestNet ? Network.TestNet : Network.Main);
+            var address = key.Derive(index).ScriptPubKey.GetDestinationAddress(_configuration.IsTestNet ? Network.TestNet : Network.Main);
             return await Task.FromResult(address.ToString());
         } 
 
-        private async Task<string> GenerateEtherumAddress(int index)
+        private async Task<string> GenerateEtherumAddress(uint index)
         {
             var wallet = new Wallet(_configuration.Mnemonic, _configuration.Passwords[Currency.ETH]);
-            return await Task.FromResult(wallet.GetAccount(index).Address);
+            return await Task.FromResult(wallet.GetAccount((int)index).Address);
         }
 
         public class Configuration
